feat: refactor & responsive for repo & repo builds
