+++ src/components/pages/repo/build-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './build-list';
