+++ src/components/pages/repo/new-build-form/index.js
@@ -1,0 +1,1 @@
+export { default } from './new-build-form';
