+++ src/components/pages/repo/new-build-form/new-build-form.module.scss
@@ -1,0 +1,58 @@
+.new-build-form {
+  display: flex;
+  flex-flow: column;
+  &-column {
+    display: flex;
+    flex-flow: column;
+    &-list {
+      span {
+        margin-bottom: 12px;
+      }
+    }
+  }
+  &-parameters {
+    display: flex;
+    & > * + * {
+      margin-left: 10px;
+    }
+    button {
+      color: #e43326;
+    }
+    &-list {
+      margin: 6px 0 12px;
+      & > * + * {
+        margin-top: 5px;
+      }
+    }
+    &-fields {
+      display: flex;
+      button {
+        color: #0092e4;
+      }
+      & > * + * {
+        margin-left: 10px;
+      }
+    }
+  }
+  &-controls {
+    padding-top: 15px;
+    & > * + * {
+      margin-left: 10px;
+    }
+    button {
+      padding: 6px 26px;
+      border-radius: 8px;
+      box-shadow: 0px 0px 1px rgba(40, 41, 61, 0.04), 0px 2px 4px rgba(96, 97, 112, 0.16);
+      box-sizing: border-box;
+      border: 1px solid #0278d5;
+    }
+    button:first-child {
+      background-color: #0278d5;
+      color: white;
+    }
+    button:last-child {
+      background: #fff;
+      color: #0278d5;
+    }
+  }
+}
