+++ src/pages/home/home.module.scss
@@ -154,8 +154,6 @@
     padding: 0;
     width: 100%;
     input {
-      // @TODO: should be removed once shared Input
-      // styles will be adjusted for responsive
       width: 100% !important;
     }
   }
