+++ src/pages/repo/repo.module.scss
@@ -1,8 +1,7 @@
+@import "styles/_variables";
+@import "styles/mixins";
+
 .header {
-  // display: flex;
-  // align-items: center;
-  // justify-content: space-between;
-  // height: 200px;
   padding: 0px 60px;
   height: 106px;
   display: flex;
@@ -16,6 +15,13 @@
   .breadcrumb {
     margin-bottom: 10px;
   }
+  @include sm-down {
+    flex-direction: column;
+    height: 180px;
+    justify-content: center;
+    align-items: flex-start;
+    padding: 0px 32px;
+  }
 }
 
 .tabs {
@@ -24,6 +30,10 @@
   filter: drop-shadow(0px 4px 4px rgba(207, 207, 207, 0.25));
   background-color: #fff;
   padding: 0px 60px;
+  @include sm-down {
+    padding: 0px 32px;
+    overflow-x: scroll;
+  }
   .tab {
     text-decoration: none;
     font-size: 14px;
@@ -79,79 +89,6 @@
   }
 
   span {
-    font-size: 13px !important;  // TODO fix base style so we can remove
-  }
-}
-
-.new-build-form {
-  display: flex;
-  flex-flow: column;
-  &-column {
-    display: flex;
-    flex-flow: column;
-    &-list {
-      span {
-        margin-bottom: 12px;
-      }
-    }
-  }
-  &-parameters {
-    display: flex;
-    & > * + * {
-      margin-left: 10px;
-    }
-    button {
-      color: #e43326;
-    }
-    &-list {
-      margin: 6px 0 12px;
-      & > * + * {
-        margin-top: 5px;
-      }
-    }
-    &-fields {
-      display: flex;
-      button {
-        color: #0092e4;
-      }
-      & > * + * {
-        margin-left: 10px;
-      }
-    }
-  }
-  &-controls {
-    padding-top: 15px;
-    & > * + * {
-      margin-left: 10px;
-    }
-    button {
-      padding: 6px 26px;
-      border-radius: 8px;
-      box-shadow: 0px 0px 1px rgba(40, 41, 61, 0.04), 0px 2px 4px rgba(96, 97, 112, 0.16);
-      box-sizing: border-box;
-      border: 1px solid #0278D5;
-    }
-    button:first-child {
-      background-color: #0278d5;
-      color: white;
-    }
-    button:last-child {
-      background: #fff;
-      color: #0278D5;
-    }
-  }
-}
-
-@media only screen and (max-width: 768px) {
-  .header {
-    flex-direction: column;
-    height: 180px;
-    justify-content: center;
-    align-items: flex-start;
-    padding: 0px 32px;
-  }
-
-  .tabs {
-    padding: 0px 32px;
+    font-size: 13px !important; // TODO fix base style so we can remove
   }
 }
